+++ src/mach/templates/partials/aws.tf
@@ -6,6 +6,11 @@ provider "aws" {
     role_arn = "arn:aws:iam::{{ aws.account_id }}:role/{{ aws.deploy_role_name }}"
   }
   {% endif %}
+  {% if aws.default_tags %}
+  default_tags {
+      tags = {{ aws.default_tags|tf }}
+  }
+  {% endif %}
 }
 
 {% for provider in aws.extra_providers %}
@@ -16,6 +21,15 @@ provider "aws" {
   {% if aws.deploy_role_name %}
   assume_role {
     role_arn = "arn:aws:iam::{{ aws.account_id }}:role/{{ aws.deploy_role_name }}"
+  }
+  {% endif %}
+  {% if provider.default_tags %}
+  default_tags {
+      tags = {{ provider.default_tags|tf }}
+  }
+  {% elif aws.default_tags %}
+  default_tags {
+      tags = {{ aws.default_tags|tf }}
   }
   {% endif %}
 }
@@ -29,6 +43,11 @@ provider "aws" {
   {% if aws.deploy_role_name %}
   assume_role {
     role_arn = "arn:aws:iam::{{ aws.account_id }}:role/{{ aws.deploy_role_name }}"
+  }
+  {% endif %}
+  {% if aws.default_tags %}
+  default_tags {
+      tags = {{ aws.default_tags|tf }}
   }
   {% endif %}
 }